requirements.txt
pycryptodome
colorama